+++ app/_components/hero.tsx
