+++ app/_components/index.ts
@@ -1,0 +1,6 @@
+export * from './infoSection';
+export * from './sponsors';
+export * from './speakers';
+export * from './meetups';
+export * from './staff';
+export * from './hero';
