+++ app/sponsors/_components/index.tsx
@@ -1,0 +1,4 @@
+export * from './planSection';
+export * from './sponsorSection';
+export * from './analyticsSection';
+export * from './principalSection';
