+++ app/sponsors/_components/principalSection.tsx
@@ -1,0 +1,10 @@
+export const PrincipalSection = () => (
+  <section>
+    <h1 className="text-5xl sm:text-8xl mb-12 mt-32 text-left font-bold w-4/5">Haz parte de nuestros sponsors</h1>
+    <p className="text-xl mb-8 text-left font-normal">
+      En BarranquillaJs, creemos en el poder de las conferencias impulsadas por la comunidad para avanzar en el mundo del desarrollo de software. Al
+      centrarnos en el talento local y el intercambio de conocimientos,estamos creando una plataforma para que tanto desarrolladores como cualquier
+      persona crezca profecionalmente.
+    </p>
+  </section>
+);
