+++ components/index.ts
@@ -1,5 +1,4 @@
 export * from './gradientBackground/gradientBackground';
-export * from './clientStyleHelper/clientStyleHelper';
 export * from './speakerCard/speakerCard';
 export * from './eventCard/eventCard';
 export * from './header/header';
