--- modules/index.ts
@@ -1,6 +1,0 @@
-export * from './infoSection/infoSection';
-export * from './sponsors/sponsors';
-export * from './speakers/speakers';
-export * from './meetups/meetups';
-export * from './staff/staff';
-export * from './hero/hero';
